fix(wasm): Cast u32 to u64 in set_max_ticks

+++ rust/src/interfaces/wasm_interface.rs
@@ -113,7 +113,7 @@ impl WasmLightVM {
   }
   #[wasm_bindgen(js_name = "setMaxTicks")]
   pub fn set_max_ticks(&mut self, value: u32) {
-    self.inner.max_ticks = value;
+    self.inner.max_ticks = value as u64;
   }
   #[wasm_bindgen(js_name = "setMaxStackSize")]
   pub fn set_max_stack_size(&mut self, value: usize) {
